+++ src/app/beer_garden/garden.py
@@ -28,9 +28,6 @@ from yapconf.exceptions import (
 
 import beer_garden.config as config
 import beer_garden.db.api as db
-from beer_garden.command_publishing_blocklist import (
-    publish_command_publishing_blocklist,
-)
 from beer_garden.db.mongo.models import RemoteUser
 from beer_garden.errors import ForwardException
 from beer_garden.events import publish, publish_event
@@ -697,7 +694,6 @@ def garden_sync(sync_target: str = None):
         logger.info("Processing garden sync, about to publish")
 
         publish_garden()
-        publish_command_publishing_blocklist()
 
     from beer_garden.router import route